Adapted model for macOS with M1 Pro chip and other improvements

+++ wan/distributed/fsdp.py
@@ -10,7 +10,7 @@ from torch.distributed.fsdp.wrap import lambda_auto_wrap_policy
 def shard_model(
     model,
     device_id,
-    param_dtype=torch.bfloat16,
+    param_dtype=torch.float32,
     reduce_dtype=torch.float32,
     buffer_dtype=torch.float32,
     process_group=None,
